+++ PI3FernandezTrigoAlejandro/ejercicios/ejercicio3recursivaconmemoria.h
@@ -14,9 +14,10 @@
 #include "types/math2.h"
 #include "types/types.h"
 #include "types/list.h"
+#include "types/hash_table.h"
 
 // Cabeceras de las funciones:
-long ejercicio3RecursivoConMemoria (int numero);
-long ejercicio3RecursivoConMemoriaInterno (int numero, hash_table memoria);
+long ejercicio3RecursivoConMemoria (long numero);
+long ejercicio3RecursivoConMemoriaInterno (long numero, hash_table memoria);
 
 #endif /* EJERCICIOS_EJERCICIO3RECURSIVACONMEMORIA_H_ */
